tests/test_requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
requests==2.32.4
urllib3==2.5.0
